+++ bin/prover-client/src/handlers/mod.rs
@@ -9,24 +9,15 @@ use tokio::sync::Mutex;
 use crate::{errors::ProvingTaskError, hosts, task2::TaskTracker};
 
 pub mod btc;
-use btc::BtcBlockspaceHandler;
-
 pub mod checkpoint;
-use checkpoint::CheckpointHandler;
-
 pub mod cl_agg;
-use cl_agg::ClAggHandler;
-
 pub mod cl_stf;
-use cl_stf::ClStfHandler;
-
 pub mod evm_ee;
-use evm_ee::EvmEeHandler;
-
+pub mod handler;
 pub mod l1_batch;
-use l1_batch::L1BatchHandler;
-
 pub mod utils;
+
+pub use handler::ProofHandler;
 
 pub trait ProvingOp {
     type Prover: ZkVmProver;
@@ -64,81 +55,5 @@ pub trait ProvingOp {
         // Requires making changes to the ProofKey to include Host as well
 
         Ok(())
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ProofHandler {
-    btc_blockspace_handler: BtcBlockspaceHandler,
-    l1_batch_handler: L1BatchHandler,
-    evm_ee_handler: EvmEeHandler,
-    cl_stf_handler: ClStfHandler,
-    cl_agg_handler: ClAggHandler,
-    checkpoint_handler: CheckpointHandler,
-}
-
-impl ProofHandler {
-    pub fn new(
-        btc_blockspace_handler: BtcBlockspaceHandler,
-        l1_batch_handler: L1BatchHandler,
-        evm_ee_handler: EvmEeHandler,
-        cl_stf_handler: ClStfHandler,
-        cl_agg_handler: ClAggHandler,
-        checkpoint_handler: CheckpointHandler,
-    ) -> Self {
-        Self {
-            btc_blockspace_handler,
-            l1_batch_handler,
-            evm_ee_handler,
-            cl_stf_handler,
-            cl_agg_handler,
-            checkpoint_handler,
-        }
-    }
-
-    pub async fn prove(
-        &self,
-        task_id: &ProofKey,
-        task_tracker: &ProofDb,
-    ) -> Result<(), ProvingTaskError> {
-        match task_id {
-            ProofKey::BtcBlockspace(_) => {
-                self.btc_blockspace_handler
-                    .prove(task_id, task_tracker)
-                    .await
-            }
-            ProofKey::L1Batch(_, _) => self.l1_batch_handler.prove(task_id, task_tracker).await,
-            ProofKey::EvmEeStf(_) => self.evm_ee_handler.prove(task_id, task_tracker).await,
-            ProofKey::ClStf(_) => self.cl_stf_handler.prove(task_id, task_tracker).await,
-            ProofKey::ClAgg(_, _) => self.cl_agg_handler.prove(task_id, task_tracker).await,
-            ProofKey::Checkpoint(_) => self.checkpoint_handler.prove(task_id, task_tracker).await,
-        }
-    }
-
-    pub async fn create_task(
-        &self,
-        task_tracker: Arc<Mutex<TaskTracker>>,
-        task_id: &ProofKey,
-    ) -> Result<(), ProvingTaskError> {
-        match task_id {
-            ProofKey::BtcBlockspace(_) => {
-                self.btc_blockspace_handler
-                    .create_task(task_tracker, task_id)
-                    .await
-            }
-            ProofKey::L1Batch(_, _) => {
-                self.l1_batch_handler
-                    .create_task(task_tracker, task_id)
-                    .await
-            }
-            ProofKey::EvmEeStf(_) => self.evm_ee_handler.create_task(task_tracker, task_id).await,
-            ProofKey::ClStf(_) => self.cl_stf_handler.create_task(task_tracker, task_id).await,
-            ProofKey::ClAgg(_, _) => self.cl_agg_handler.create_task(task_tracker, task_id).await,
-            ProofKey::Checkpoint(_) => {
-                self.checkpoint_handler
-                    .create_task(task_tracker, task_id)
-                    .await
-            }
-        }
     }
 }
