+++ bin/prover-client/src/main.rs
@@ -4,38 +4,25 @@ use std::sync::Arc;
 
 use args::Args;
 use config::NUM_PROVER_WORKERS;
-// use ckp_runner::start_checkpoints_task;
-// use dispatcher::TaskDispatcher;
+use handlers::ProofHandler;
 use jsonrpsee::http_client::HttpClientBuilder;
 use manager2::ProverManager;
-// use manager::ProverManager;
-// use proving_ops::{
-//     btc_ops::BtcOperations, checkpoint_ops::CheckpointOperations, cl_ops::ClOperations,
-//     el_ops::ElOperations, l1_batch_ops::L1BatchOperations, l2_batch_ops::L2BatchOperations,
-// };
-use rpc_server::{ProverClientRpc, ServerRequest};
+use rpc_server::ProverClientRpc;
 use strata_btcio::rpc::BitcoinClient;
 use strata_common::logging;
-use strata_primitives::proof::ProofKey;
 use task2::TaskTracker;
-use tokio::sync::{mpsc, Mutex};
+use tokio::sync::Mutex;
 use tracing::{debug, info};
 
 mod args;
-// mod ckp_runner;
 mod config;
 mod db;
-// mod dispatcher;
 mod errors;
 mod handlers;
 mod hosts;
-// mod manager;
 mod manager2;
 mod primitives;
-// mod prover;
-// mod proving_ops;
 mod rpc_server;
-// mod task;
 mod task2;
 
 #[tokio::main]
@@ -63,35 +50,33 @@ async fn main() {
     )
     .expect("failed to connect to the btc client");
 
-    let (task_tx, task_rx) = mpsc::channel::<ServerRequest>(64);
+    let handler = Arc::new(ProofHandler::init(btc_client, el_client, cl_client));
+    let task_tracker = Arc::new(Mutex::new(TaskTracker::new()));
 
-    let manager = ProverManager::init(
-        btc_client,
-        el_client,
-        cl_client,
-        task_rx,
-        NUM_PROVER_WORKERS,
-    );
+    let manager = ProverManager::new(task_tracker.clone(), handler.clone(), NUM_PROVER_WORKERS);
+    debug!("Initialized Prover Manager");
 
     // run prover manager in background
-    tokio::spawn(async move { manager.process().await });
+    tokio::spawn(async move { manager.process_pending_tasks().await });
+    debug!("Spawn process pending tasks");
 
-    // // Run prover manager in dev mode or runner mode
-    // if args.enable_dev_rpcs {
-    //     // Run the rpc server on dev mode only
-    //     let rpc_url = args.get_dev_rpc_url();
-    //     run_rpc_server(manager.clone(), rpc_url, args.enable_dev_rpcs)
-    //         .await
-    //         .expect("prover client rpc")
-    // }
+    // Run prover manager in dev mode or runner mode
+    if args.enable_dev_rpcs {
+        // Run the rpc server on dev mode only
+        let rpc_url = args.get_dev_rpc_url();
+        run_rpc_server(task_tracker, handler, rpc_url, args.enable_dev_rpcs)
+            .await
+            .expect("prover client rpc")
+    }
 }
 
 async fn run_rpc_server(
-    task_tx: mpsc::Sender<ServerRequest>,
+    task_tracker: Arc<Mutex<TaskTracker>>,
+    handler: Arc<ProofHandler>,
     rpc_url: String,
     enable_dev_rpc: bool,
 ) -> anyhow::Result<()> {
-    let rpc_impl = ProverClientRpc::new(task_tx);
+    let rpc_impl = ProverClientRpc::new(task_tracker, handler);
     rpc_server::start(&rpc_impl, rpc_url, enable_dev_rpc).await?;
     anyhow::Ok(())
 }
