+++ bin/prover-client/src/primitives/mod.rs
@@ -1,4 +1,2 @@
-// pub mod prover_input;
 pub mod status;
-// pub mod tasks_scheduler;
 pub mod vms;
