--- bin/prover-client/src/proving_ops/mod.rs
@@ -1,7 +1,0 @@
-pub mod btc_ops;
-pub mod checkpoint_ops;
-pub mod cl_ops;
-pub mod el_ops;
-pub mod l1_batch_ops;
-pub mod l2_batch_ops;
-pub mod ops;
