+++ crates/rpc/types/src/lib.rs
@@ -2,4 +2,5 @@ pub mod errors;
 pub mod types;
 
 pub use errors::*;
+pub use strata_primitives::proof::ProofKey;
 pub use types::*;
