+++ crates/zkvm/adapters/sp1/src/host.rs
@@ -64,6 +64,7 @@ impl ZkVmHost for SP1Host {
             std::env::set_var("SP1_PROVER", "mock");
         }
 
+        // sp1_sdk::utils::setup_logger();
         let client = ProverClient::new();
 
         // Start proving
