+++ functional-tests/run_test.sh
@@ -11,8 +11,9 @@ fi
 
 # Conditionally run cargo build based on PROVER_TEST
 if [ ! -z $PROVER_TEST ]; then
-    echo "Running on prover mode"
-    cargo build --release
+    echo "Running on prover mock mode"
+    # cargo build -F "prover"
+    cargo build -F "prover, mock" --release
 	export PATH=$(realpath ../target/release/):$PATH
 else
     echo "Running on seq mode"
